// where-in-the-world/src/types/countries.ts
export interface Country {
  name: Name;
  tld?: string[];
  cca2: string;
  ccn3?: string;
  cca3: string;
  cioc?: string;
  independent?: boolean;
  status: string;
  unMember: boolean;
  currencies?: Currencies;
  idd: Idd;
  capital?: string[];
  altSpellings: string[];
  region: string;
  subregion?: string;
  languages?: Languages;
  translations: Translations;
  latlng: number[];
  landlocked: boolean;
  borders?: string[];
  area: number;
  demonyms?: Demonyms;
  flag: string;
  maps: Maps;
  population: number;
  fifa?: string;
  car: Car;
  timezones: string[];
  continents: string[];
  flags: Flags;
  coatOfArms: CoatOfArms;
  startOfWeek: string;
  capitalInfo: CapitalInfo;
  postalCode?: PostalCode;
  gini?: Gini;
}

export interface Name {
  common: string;
  official: string;
  nativeName?: NativeName;
}

export interface NativeName {
  crs?: Crs;
  eng?: Eng;
  fra?: Fra;
  spa?: Spa;
  smo?: Smo;
  por?: Por;
  urd?: Urd;
  nld?: Nld;
  swa?: Swa;
  ron?: Ron;
  rus?: Rus;
  uzb?: Uzb;
  mkd?: Mkd;
  vie?: Vie;
  nrf?: Nrf;
  ara?: Ara;
  tha?: Tha;
  afr?: Afr;
  deu?: Deu;
  her?: Her;
  hgm?: Hgm;
  kwn?: Kwn;
  loz?: Loz;
  ndo?: Ndo;
  tsn?: Tsn;
  kat?: Kat;
  tuk?: Tuk;
  kaz?: Kaz;
  pap?: Pap;
  dan?: Dan;
  fao?: Fao;
  nau?: Nau;
  zho?: Zho;
  jam?: Jam;
  lav?: Lav;
  fil?: Fil;
  nbl?: Nbl;
  nso?: Nso;
  sot?: Sot;
  ssw?: Ssw;
  tso?: Tso;
  ven?: Ven;
  xho?: Xho;
  zul?: Zul;
  jpn?: Jpn;
  mon?: Mon;
  hye?: Hye;
  slk?: Slk;
  msa?: Msa;
  ind?: Ind;
  gsw?: Gsw;
  ita?: Ita;
  roh?: Roh;
  kal?: Kal;
  kon?: Kon;
  lin?: Lin;
  lua?: Lua;
  grn?: Grn;
  ltz?: Ltz;
  tkl?: Tkl;
  hun?: Hun;
  nor?: Nor;
  mlg?: Mlg;
  hmo?: Hmo;
  tpi?: Tpi;
  sag?: Sag;
  aze?: Aze;
  zdj?: Zdj;
  srp?: Srp;
  nfr?: Nfr;
  gle?: Gle;
  nep?: Nep;
  dzo?: Dzo;
  ukr?: Ukr;
  fij?: Fij;
  hif?: Hif;
  ell?: Ell;
  tur?: Tur;
  heb?: Heb;
  bwg?: Bwg;
  kck?: Kck;
  khi?: Khi;
  ndc?: Ndc;
  nde?: Nde;
  nya?: Nya;
  sna?: Sna;
  toi?: Toi;
  zib?: Zib;
  hat?: Hat;
  pau?: Pau;
  lat?: Lat;
  cha?: Cha;
  bos?: Bos;
  hrv?: Hrv;
  amh?: Amh;
  swe?: Swe;
  pol?: Pol;
  cat?: Cat;
  lao?: Lao;
  bis?: Bis;
  ces?: Ces;
  isl?: Isl;
  nno?: Nno;
  nob?: Nob;
  smi?: Smi;
  cnr?: Cnr;
  sin?: Sin;
  tam?: Tam;
  ben?: Ben;
  kin?: Kin;
  ber?: Ber;
  arc?: Arc;
  ckb?: Ckb;
  glv?: Glv;
  kor?: Kor;
  fas?: Fas;
  sqi?: Sqi;
  tgk?: Tgk;
  aym?: Aym;
  que?: Que;
  bar?: Bar;
  bjz?: Bjz;
  pov?: Pov;
  mah?: Mah;
  mya?: Mya;
  som?: Som;
  prs?: Prs;
  pus?: Pus;
  rar?: Rar;
  mey?: Mey;
  mri?: Mri;
  nzs?: Nzs;
  tir?: Tir;
  khm?: Khm;
  bel?: Bel;
  pih?: Pih;
  tvl?: Tvl;
  bul?: Bul;
  niu?: Niu;
  est?: Est;
  mlt?: Mlt;
  slv?: Slv;
  hin?: Hin;
  run?: Run;
  lit?: Lit;
  ton?: Ton;
  cal?: Cal;
  mfe?: Mfe;
  gil?: Gil;
  tet?: Tet;
  div?: Div;
  kir?: Kir;
  fin?: Fin;
}
// ...
export interface Currencies {
  SCR?: Scr;
  USD?: Usd;
  CRC?: Crc;
  VES?: Ves;
  AOA?: Aoa;
  PAB?: Pab;
  PKR?: Pkr;
  XOF?: Xof;
  EUR?: Eur;
  XAF?: Xaf;
  ZMW?: Zmw;
  UGX?: Ugx;
  RON?: Ron2;
  UZS?: Uzs;
  MKD?: Mkd2;
  MDL?: Mdl;
  GTQ?: Gtq;
  VND?: Vnd;
  GBP?: Gbp;
  JEP?: Jep;
  FKP?: Fkp;
  BHD?: Bhd;
  THB?: Thb;
  SSP?: Ssp;
  NAD?: Nad;
  ZAR?: Zar;
  GEL?: Gel;
  WST?: Wst;
  SLL?: Sll;
  TMT?: Tmt;
  OMR?: Omr;
  KZT?: Kzt;
  AWG?: Awg;
  TND?: Tnd;
  MXN?: Mxn;
  LRD?: Lrd;
  AUD?: Aud;
  RUB?: Rub;
  SDG?: Sdg;
  GYD?: Gyd;
  DKK?: Dkk;
  FOK?: Fok;
  MOP?: Mop;
  JMD?: Jmd;
  PHP?: Php;
  JPY?: Jpy;
  MNT?: Mnt;
  AMD?: Amd;
  KYD?: Kyd;
  MYR?: Myr;
  LBP?: Lbp;
  IDR?: Idr;
  CHF?: Chf;
  JOD?: Jod;
  CDF?: Cdf;
  ARS?: Ars;
  GMD?: Gmd;
  TTD?: Ttd;
  NZD?: Nzd;
  HUF?: Huf;
  MGA?: Mga;
  PGK?: Pgk;
  XPF?: Xpf;
  DJF?: Djf;
  NOK?: Nok;
  TWD?: Twd;
  EGP?: Egp;
  ILS?: Ils;
  AZN?: Azn;
  KMF?: Kmf;
  GHS?: Ghs;
  BWP?: Bwp;
  RSD?: Rsd;
  DOP?: Dop;
  XCD?: Xcd;
  GGP?: Ggp;
  AED?: Aed;
  NPR?: Npr;
  BTN?: Btn;
  INR?: Inr;
  HKD?: Hkd;
  UAH?: Uah;
  FJD?: Fjd;
  ANG?: Ang;
  ZWL?: Zwl;
  SZL?: Szl;
  YER?: Yer;
  HTG?: Htg;
  CVE?: Cve;
  UYU?: Uyu;
  BAM?: Bam;
  ETB?: Etb;
  PLN?: Pln;
  CUC?: Cuc;
  CUP?: Cup;
  KES?: Kes;
  LAK?: Lak;
  CAD?: Cad;
  NGN?: Ngn;
  VUV?: Vuv;
  CZK?: Czk;
  MWK?: Mwk;
  ISK?: Isk;
  CLP?: Clp;
  BMD?: Bmd;
  KWD?: Kwd;
  LKR?: Lkr;
  CNY?: Cny;
  BDT?: Bdt;
  SEK?: Sek;
  TRY?: Try;
  GNF?: Gnf;
  TZS?: Tzs;
  RWF?: Rwf;
  SGD?: Sgd;
  MAD?: Mad;
  IQD?: Iqd;
  BND?: Bnd;
  IMP?: Imp;
  KPW?: Kpw;
  IRR?: Irr;
  PYG?: Pyg;
  ALL?: All;
  TJS?: Tjs;
  BOB?: Bob;
  COP?: Cop;
  BZD?: Bzd;
  MMK?: Mmk;
  BRL?: Brl;
  SOS?: Sos;
  AFN?: Afn;
  CKD?: Ckd;
  DZD?: Dzd;
  MRU?: Mru;
  ERN?: Ern;
  KHR?: Khr;
  BSD?: Bsd;
  BYN?: Byn;
  TVD?: Tvd;
  SHP?: Shp;
  BGN?: Bgn;
  MZN?: Mzn;
  PEN?: Pen;
  BIF?: Bif;
  HNL?: Hnl;
  TOP?: Top;
  SAR?: Sar;
  SRD?: Srd;
  QAR?: Qar;
  GIP?: Gip;
  MUR?: Mur;
  BBD?: Bbd;
  SYP?: Syp;
  STN?: Stn;
  KID?: Kid;
  LSL?: Lsl;
  SBD?: Sbd;
  LYD?: Lyd;
  KRW?: Krw;
  NIO?: Nio;
  MVR?: Mvr;
  KGS?: Kgs;
}

export interface Scr {
  name: string;
  symbol: string;
}

export interface Usd {
  name: string;
  symbol: string;
}

export interface Crc {
  name: string;
  symbol: string;
}

export interface Ves {
  name: string;
  symbol: string;
}

export interface Aoa {
  name: string;
  symbol: string;
}

export interface Pab {
  name: string;
  symbol: string;
}

export interface Pkr {
  name: string;
  symbol: string;
}

export interface Xof {
  name: string;
  symbol: string;
}

export interface Eur {
  name: string;
  symbol: string;
}

export interface Xaf {
  name: string;
  symbol: string;
}

export interface Zmw {
  name: string;
  symbol: string;
}

export interface Ugx {
  name: string;
  symbol: string;
}

export interface Ron2 {
  name: string;
  symbol: string;
}

export interface Uzs {
  name: string;
  symbol: string;
}

export interface Mkd2 {
  name: string;
  symbol: string;
}

export interface Mdl {
  name: string;
  symbol: string;
}

export interface Gtq {
  name: string;
  symbol: string;
}

export interface Vnd {
  name: string;
  symbol: string;
}

export interface Gbp {
  name: string;
  symbol: string;
}

export interface Jep {
  name: string;
  symbol: string;
}

export interface Fkp {
  name: string;
  symbol: string;
}

export interface Bhd {
  name: string;
  symbol: string;
}

export interface Thb {
  name: string;
  symbol: string;
}

export interface Ssp {
  name: string;
  symbol: string;
}

export interface Nad {
  name: string;
  symbol: string;
}

export interface Zar {
  name: string;
  symbol: string;
}

export interface Gel {
  name: string;
  symbol: string;
}

export interface Wst {
  name: string;
  symbol: string;
}

export interface Sll {
  name: string;
  symbol: string;
}

export interface Tmt {
  name: string;
  symbol: string;
}

export interface Omr {
  name: string;
  symbol: string;
}

export interface Kzt {
  name: string;
  symbol: string;
}

export interface Awg {
  name: string;
  symbol: string;
}

export interface Tnd {
  name: string;
  symbol: string;
}

export interface Mxn {
  name: string;
  symbol: string;
}

export interface Lrd {
  name: string;
  symbol: string;
}

export interface Aud {
  name: string;
  symbol: string;
}

export interface Rub {
  name: string;
  symbol: string;
}

export interface Sdg {
  name: string;
}

export interface Gyd {
  name: string;
  symbol: string;
}

export interface Dkk {
  name: string;
  symbol: string;
}

export interface Fok {
  name: string;
  symbol: string;
}

export interface Mop {
  name: string;
  symbol: string;
}

export interface Jmd {
  name: string;
  symbol: string;
}

export interface Php {
  name: string;
  symbol: string;
}

export interface Jpy {
  name: string;
  symbol: string;
}

export interface Mnt {
  name: string;
  symbol: string;
}

export interface Amd {
  name: string;
  symbol: string;
}

export interface Kyd {
  name: string;
  symbol: string;
}

export interface Myr {
  name: string;
  symbol: string;
}

export interface Lbp {
  name: string;
  symbol: string;
}

export interface Idr {
  name: string;
  symbol: string;
}

export interface Chf {
  name: string;
  symbol: string;
}

export interface Jod {
  name: string;
  symbol: string;
}

export interface Cdf {
  name: string;
  symbol: string;
}

export interface Ars {
  name: string;
  symbol: string;
}
// ...
export interface Idd {
  root?: string;
  suffixes?: string[];
}

export interface Languages {
  crs?: string;
  eng?: string;
  fra?: string;
  spa?: string;
  smo?: string;
  por?: string;
  urd?: string;
  nld?: string;
  cat?: string;
  eus?: string;
  glc?: string;
  swa?: string;
  ron?: string;
  rus?: string;
  uzb?: string;
  mkd?: string;
  vie?: string;
  nrf?: string;
  ara?: string;
  tha?: string;
  afr?: string;
  deu?: string;
  her?: string;
  hgm?: string;
  kwn?: string;
  loz?: string;
  ndo?: string;
  tsn?: string;
  kat?: string;
  tuk?: string;
  kaz?: string;
  pap?: string;
  dan?: string;
  fao?: string;
  nau?: string;
  zho?: string;
  jam?: string;
  lav?: string;
  fil?: string;
  nbl?: string;
  nso?: string;
  sot?: string;
  ssw?: string;
  tso?: string;
  ven?: string;
  xho?: string;
  zul?: string;
  jpn?: string;
  mon?: string;
  hye?: string;
  slk?: string;
  msa?: string;
  ind?: string;
  gsw?: string;
  ita?: string;
  roh?: string;
  kal?: string;
  kon?: string;
  lin?: string;
  lua?: string;
  grn?: string;
  ltz?: string;
  tkl?: string;
  hun?: string;
  nor?: string;
  mlg?: string;
  hmo?: string;
  tpi?: string;
  sag?: string;
  aze?: string;
  zdj?: string;
  srp?: string;
  nfr?: string;
  gle?: string;
  nep?: string;
  dzo?: string;
  ukr?: string;
  fij?: string;
  hif?: string;
  ell?: string;
  tur?: string;
  heb?: string;
  bwg?: string;
  kck?: string;
  khi?: string;
  ndc?: string;
  nde?: string;
  nya?: string;
  sna?: string;
  toi?: string;
  zib?: string;
  hat?: string;
  pau?: string;
  lat?: string;
  cha?: string;
  bos?: string;
  hrv?: string;
  amh?: string;
  swe?: string;
  pol?: string;
  lao?: string;
  bis?: string;
  ces?: string;
  isl?: string;
  nno?: string;
  nob?: string;
  smi?: string;
  cnr?: string;
  sin?: string;
  tam?: string;
  ben?: string;
  kin?: string;
  ber?: string;
  arc?: string;
  ckb?: string;
  glv?: string;
  kor?: string;
  fas?: string;
  sqi?: string;
  tgk?: string;
  aym?: string;
  que?: string;
  de?: string;
  bjz?: string;
  pov?: string;
  mah?: string;
  mya?: string;
  som?: string;
  prs?: string;
  pus?: string;
  rar?: string;
  mey?: string;
  mri?: string;
  nzs?: string;
  tir?: string;
  khm?: string;
  bel?: string;
  pih?: string;
  tvl?: string;
  bul?: string;
  niu?: string;
  est?: string;
  mlt?: string;
  slv?: string;
  hin?: string;
  run?: string;
  lit?: string;
  ton?: string;
  cal?: string;
  mfe?: string;
  gil?: string;
  tet?: string;
  div?: string;
  kir?: string;
  fin?: string;
}

export interface Translations {
  ara: Ara2;
  bre: Bre;
  ces: Ces2;
  cym: Cym;
  deu: Deu2;
  est: Est2;
  fin: Fin2;
  fra: Fra2;
  hrv?: Hrv2;
  hun: Hun2;
  ita: Ita2;
  jpn?: Jpn2;
  kor: Kor2;
  nld: Nld2;
  per?: Per;
  pol: Pol2;
  por: Por2;
  rus: Rus2;
  slk: Slk2;
  spa: Spa2;
  srp: Srp2;
  swe: Swe2;
  tur: Tur2;
  urd: Urd2;
  zho?: Zho2;
}
// ...
export interface Demonyms {
  eng: Eng2;
  fra?: Fra3;
}

export interface Eng2 {
  f: string;
  m: string;
}

export interface Fra3 {
  f: string;
  m: string;
}

export interface Maps {
  googleMaps: string;
  openStreetMaps: string;
}

export interface Car {
  signs?: string[];
  side: string;
}

export interface Flags {
  png: string;
  svg: string;
  alt?: string;
}

export interface CoatOfArms {
  png?: string;
  svg?: string;
}

export interface CapitalInfo {
  latlng?: number[];
}

export interface PostalCode {
  format: string;
  regex?: string;
}

export interface Gini {
  "2018"?: number;
  "2019"?: number;
  "2006"?: number;
  "2015"?: number;
  "2017"?: number;
  "2016"?: number;
  "2003"?: number;
  "2014"?: number;
  "2011"?: number;
  "2013"?: number;
  "1998"?: number;
  "2012"?: number;
  "2004"?: number;
  "2010"?: number;
  "1992"?: number;
  "2009"?: number;
  "2008"?: number;
  "2005"?: number;
  "1999"?: number;
}